fix: fix wide SVGs being cut at 300px
Fixes wide SVGs (e.g. gitgraphs) from being cut at 300px.

When including an SVG as in image, most browsers will try
to find the width of the SVG from the SVG file.

However, if a browser sees `width=100%`, it will normally
default to 300px, truncating wide SVGs.

To fix this, this commit uses SVGDom to calcalute the actual
width of the SVG in pixels, before inserting it into the dataurl.

Because SVGDom and the browser's SVG renderer may be slighlty
different, the calculated widths may be slightly different too,
but I can't figure out a better solution.