fix: throw error if mermaid does not produce SVG
An error is now thrown if mermaid-cli fails to render an SVG.

Previously, the `Promise` returned by `renderMermaidFile` would stall
forever if `mermaid-cli` exited without an error code, and without
rendering an SVG.

This is because mermaid-cli does not handle Promise rejections,
so if there is a Promise rejection, the exit code of mermaid-cli
will be `0` (e.g. success).