fix: try to automatically fix invalid SVGs
Tries to automatically fix invalid SVGs returned by `mermaid-cli`.

Somtimes `mermaid-cli` returns an SVG wrapped in a `<div>`,
which caused a confusing error to occur.

For example, the "Syntax error in graph" SVG created when
there is an error in your mermaid diagram is returned
by `mermaid-cli` as an SVG wrapped in a `<div>`.